fix(attributeController): enhance error handling and validation messages for attribute operations

+++ src/controller/attributeController.js
@@ -1,20 +1,22 @@
 const Attribute = require("../models/Attribute");
+const ScoringFormula = require("../models/ScoringFormula");
 
 const attributeController = {
-  // Tạo mới một Attribute
   createAttribute: async (req, res) => {
     try {
       const { name } = req.body;
       const existingAttribute = await Attribute.findOne({ name });
       if (existingAttribute) {
-        return res
-          .status(400)
-          .json({ message: "Attribute name already exists" });
+        return res.status(400).json({ message: "Thuộc tính đã tồn tại" });
       }
       const attribute = new Attribute(req.body);
       await attribute.save();
       res.status(201).json(attribute);
     } catch (error) {
+      if (error.name === "ValidationError") {
+        const errors = Object.values(error.errors).map((err) => err.message);
+        return res.status(400).json({ message: "Lỗi xác thực", errors });
+      }
       res.status(400).json({ message: error.message });
     }
   },
@@ -67,7 +69,19 @@ const attributeController = {
   // Xóa Attribute theo tên
   deleteAttributeByName: async (req, res) => {
     try {
-      const { name } = req.body;
+      const { name } = req.params;
+
+      // Kiểm tra xem thuộc tính có đang được dùng trong công thức điểm không
+      const formulaUsingAttribute = await ScoringFormula.findOne({
+        attributeName: name,
+      });
+      if (formulaUsingAttribute) {
+        return res.status(400).json({
+          message:
+            "Không thể xóa thuộc tính vì đang được sử dụng trong công thức điểm.",
+        });
+      }
+
       const attribute = await Attribute.findOneAndDelete({ name });
       if (!attribute) {
         return res.status(404).json({ message: "Attribute not found" });
